\name{GGfitter-class}
\docType{class}
\alias{GGfitter-class}

\title{Class "GGfitter" legacy container for
fast fitting of eQTL search models -- obsolete, see
gwSnpScreen for more current approaches}
\description{Class "GGfitter" legacy container for
fast fitting of eQTL search models -- obsolete, see
gwSnpScreen for more current approaches}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("GGfitter", ...)}.
}
\section{Slots}{
	 \describe{
    \item{\code{name}:}{Object of class \code{"character"} reflectance support }
    \item{\code{func}:}{Object of class \code{"function"} actual fitting
function }
  }
}
\section{Methods}{
No methods defined with class "GGfitter" in the signature.
Use with snpScreen, but see gwSnpScreen for more current approach.
}
%\references{ ~put references to the literature/web site here ~ }
%\author{ ~~who you are~~ }
%\note{ ~~further notes~~ }
%
% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%	~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%	or \code{\linkS4class{CLASSNAME}} for links to other classes
%}
\examples{
showClass("GGfitter")
}
\keyword{classes}
